functions

TheCannon.expanded_size — Function

expandeded_size(nlabels; quadratic=true)

The length of a label vector of length nlabels after it has been quadratically expanded.

See also: collapsed_size

TheCannon.collapsed_size — Function

collapsed_size(nelabels; quadratic=true)

The length of a label vector corresponding to an expanded label vector of length nelabels.

See also: expanded_size

TheCannon.expand_labels — Function

expand_labels(labels; quadratic=true)

If labels is a Vector, return it's quadratic (or linear) expansion. If labels is a Matrix, return the matrix whose rows are expansions of the rows of labels.

This is the transformation referred to as $\eta$ in Wheeler+ 2020, and as the "vectorizing function" in Casey+ 2016.

TheCannon.standardize_labels — Function

Standardize the label matrix (nstars x nlabels) to live roughly in [-1, 1] by subtractive the mean and dividing by the scatter. returns (standardized_labels, pivots, scales)

TheCannon.unstandardize_labels — Function

Transform labels back to their unstandardize form. Returns

labels.*transpose(hcat(scale)) .+ transpose(hcat(pivot))

TheCannon.train — Function

train(flux, ivar, labels, mask=nothing, verbose=true, quadratic=true)

returns: theta, scatters Run the training step of The Cannon, i.e. calculate coefficients for each pixel.

flux contains the spectra for each star in the training set. It should be nstars x npixels (row-vectors are spectra)

ivar contains the inverse variance for each pixel in the same shape as flux

labels contains the labels for each star. It should be nstars x nlabels. It will be expanded into the quadratic label space before training.

mask (optional) is a nlabels x npix matrix of booleans specifying which labels are "allowed" to affect the spectrum at each pixel.

returns:

theta: the matrix containing the cannon coefficients. It will be n_expanded_labels x npix

scatters: the model scatter at each pixel

TheCannon.infer — Function

infer(flux, ivar, theta, scatters; quadratic=true, verbose=true)

Run the test step of the cannon. Given theta and scatters (from training), infer stellar parameters.

flux contains the spectra for each star for which you want to infer labels

in the training set. It should be nstars x npixels (row-vectors are spectra)

ivar contains the inverse variance for each pixel in the same shape as flux

theta: the matrix containing the cannon coefficients. It will be n_expanded_labels x npix

scatters: the model scatter at each pixel

If verbose is set to true, gives an update for every 100 stars processed.

TheCannon.quad_coeff_matrix — Function

Get the quadratic terms of theta as symetric matrices. returns an array of dimensions nlabels x nlabels x npixels

Q = quad_coeff_matrix(theta)
Q[:, :, 1] #quadratic coefficients for first pixel