-

functions

TheCannon.expanded_size — Function.

expandeded_size(nlabels; quadratic=true)

The length of a label vector of length nlabels after it has been quadratically expanded.

See also: collapsed_size

source

TheCannon.collapsed_size — Function.

collapsed_size(nelabels; quadratic=true)

The length of a label vector corresponding to an expanded label vector of length nelabels.

See also: expanded_size

source

Missing docstring.

Missing docstring for expand_labels. Check Documenter's build log for details.

Missing docstring.

Missing docstring for standardize_labels. Check Documenter's build log for details.

Missing docstring.

Missing docstring for unstandardize_labels. Check Documenter's build log for details.

TheCannon.train — Function.

train(flux, ivar, labels)

returns: theta, scatters Run the training step of The Cannon, i.e. calculate coefficients for each pixel.

flux contains the spectra for each pixel in the training set. It should be nstars x npixels (row-vectors are spectra)

ivar contains the inverse variance for each pixel in the same shape as flux

labels contains the labels for each star. It should be nstars x nlabels. It will be expanded into the quadratic label space before training.

source

TheCannon.infer — Function.

infer(flux, ivar, theta, scatters)

Run the test step of the cannon. Given a Cannon model (from training), infer stellar parameters

source

TheCannon.quad_coeff_matrix — Function.

Get the quadratic terms of theta as matrices. returns an array of dimensions nlabels x nlabels x npixels

Q = quadcoeffmatrix(theta) Q[:, :, 1] #quadratic coefficients for first pixel

source