-

functions

TheCannon.expanded_size — Function.

expandeded_size(nlabels; quadratic=true)

The length of a label vector of length nlabels after it has been quadratically expanded.

See also: collapsed_size

source

TheCannon.collapsed_size — Function.

collapsed_size(nelabels; quadratic=true)

The length of a label vector corresponding to an expanded label vector of length nelabels.

See also: expanded_size

source

Missing docstring.

Missing docstring for expand_labels. Check Documenter's build log for details.

TheCannon.standardize_labels — Function.

Transform the label matrix (nstars x nlabels) live roughly in [-1, 1] by subtractive the mean and dividing by the scatter. returns (standardized_labels, pivots, scales)

source

TheCannon.unstandardize_labels — Function.

Transform labels back to their original scale.

source

TheCannon.train — Function.

train(flux, ivar, labels)

returns: theta, scatters Run the training step of The Cannon, i.e. calculate coefficients for each pixel.

flux contains the spectra for each star in the training set. It should be nstars x npixels (row-vectors are spectra)

ivar contains the inverse variance for each pixel in the same shape as flux

labels contains the labels for each star. It should be nstars x nlabels. It will be expanded into the quadratic label space before training.

returns:

theta: the matrix containing the cannon coefficients. It will be n_expanded_labels x npix

scatters: the model scatter at each pixel

source

TheCannon.infer — Function.

infer(flux, ivar, theta, scatters)

Run the test step of the cannon. Given theta and scatters (from training), infer stellar parameters.

flux contains the spectra for each star for which you want to infer labels

in the training set. It should be nstars x npixels (row-vectors are spectra)

ivar contains the inverse variance for each pixel in the same shape as flux

theta: the matrix containing the cannon coefficients. It will be n_expanded_labels x npix

scatters: the model scatter at each pixel

source

TheCannon.quad_coeff_matrix — Function.

Get the quadratic terms of theta as matrices. returns an array of dimensions nlabels x nlabels x npixels ``` Q = quadcoeffmatrix(theta) Q[:, :, 1] #quadratic coefficients for first pixel 111

source